# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/admin/info": {
            "get": {
                "description": "If both params are supplied, ID is used and request ID is ignored",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin",
                    "info"
                ],
                "summary": "Finds a request info by ID or request ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "request ID",
                        "name": "requestId",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.RequestInfoDTO"
                        }
                    },
                    "400": {
                        "description": "invalid request parameters",
                        "schema": {
                            "$ref": "#/definitions/v1.genericErrorResponse"
                        }
                    },
                    "404": {
                        "description": "request info not found",
                        "schema": {
                            "$ref": "#/definitions/v1.genericErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.genericErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/admin/link": {
            "post": {
                "description": "Returns a shortened link for the given URL",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin",
                    "links"
                ],
                "summary": "Create a short link",
                "parameters": [
                    {
                        "description": "Request body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.createShortLinkRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.shortLinkDto"
                        }
                    },
                    "400": {
                        "description": "slug already used",
                        "schema": {
                            "$ref": "#/definitions/v1.genericErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.genericErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/shorten": {
            "post": {
                "description": "Returns a shortened link for the given URL",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Shorten an URL",
                "parameters": [
                    {
                        "description": "Request body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.shortenUrlRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.shortenUrlResponse"
                        }
                    },
                    "400": {
                        "description": "slug already used",
                        "schema": {
                            "$ref": "#/definitions/v1.genericErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.genericErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.RequestInfoDTO": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "headers": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "method": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "query": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "realIP": {
                    "type": "string"
                },
                "requestId": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                },
                "userAgent": {
                    "type": "string"
                }
            }
        },
        "v1.createShortLinkRequest": {
            "type": "object",
            "required": [
                "originalUrl"
            ],
            "properties": {
                "expiresAt": {
                    "description": "needs to be RFC3339 with timezone (or UTC)",
                    "type": "string"
                },
                "originalUrl": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                }
            }
        },
        "v1.genericErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "v1.shortLinkDto": {
            "type": "object",
            "properties": {
                "expiresAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "originalUrl": {
                    "type": "string"
                },
                "shortUrl": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                }
            }
        },
        "v1.shortenUrlRequest": {
            "type": "object",
            "required": [
                "originalUrl"
            ],
            "properties": {
                "expiresAt": {
                    "description": "needs to be RFC3339 with timezone (or UTC)",
                    "type": "string"
                },
                "originalUrl": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                }
            }
        },
        "v1.shortenUrlResponse": {
            "type": "object",
            "properties": {
                "shortUrl": {
                    "type": "string"
                }
            }
        },
...
    }
}